[core] Adding React-Native SHA256  platforms
