+++ sdk/core/core-util/.tshy/browser.json
@@ -6,6 +6,7 @@
     "../src/**/*.tsx"
   ],
   "exclude": [
+    ".././src/bytesEncoding-react-native.mts",
     ".././src/sha256-react-native.mts",
     ".././src/uuidUtils-react-native.mts"
   ],
