+++ sdk/core/core-util/.tshy/esm.json
@@ -9,6 +9,7 @@
     ".././src/bytesEncoding-browser.mts",
     ".././src/sha256-browser.mts",
     ".././src/uuidUtils-browser.mts",
+    ".././src/bytesEncoding-react-native.mts",
     ".././src/sha256-react-native.mts",
     ".././src/uuidUtils-react-native.mts"
   ],
