+++ sdk/core/core-util/src/bytesEncoding-react-native.mts
