+++ sdk/core/core-util/src/sha256-browser.mts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export { computeSha256Hash } from "./sha256.common.js";
+export * from "./sha256.common.js";
