+++ sdk/core/ts-http-runtime/.tshy/browser.json
@@ -8,6 +8,7 @@
   "exclude": [
     ".././src/defaultHttpClient-react-native.mts",
     ".././src/logger/log-react-native.mts",
+    ".././src/util/bytesEncoding-react-native.mts",
     ".././src/util/sha256-react-native.mts",
     ".././src/util/userAgentPlatform-react-native.mts",
     ".././src/util/uuidUtils-react-native.mts"
