+++ sdk/core/ts-http-runtime/.tshy/commonjs.json
@@ -19,6 +19,7 @@
     "../src/util/uuidUtils-browser.mts",
     "../src/defaultHttpClient-react-native.mts",
     "../src/logger/log-react-native.mts",
+    "../src/util/bytesEncoding-react-native.mts",
     "../src/util/sha256-react-native.mts",
     "../src/util/userAgentPlatform-react-native.mts",
     "../src/util/uuidUtils-react-native.mts"
