+++ sdk/core/ts-http-runtime/src/util/bytesEncoding-react-native.mts
